Move partitionArray to a util module
